+++ samples/MediatR.Extensions.Autofac.DepdencyInjection.Samples.WebApi/Program.cs
@@ -1,9 +1,11 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Autofac.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace MediatR.Extensions.Autofac.DepdencyInjection.Samples.WebApi
 {
+    // ReSharper disable once ClassNeverInstantiated.Global
     public class Program
     {
         public static void Main(string[] args)
@@ -11,9 +13,9 @@ namespace MediatR.Extensions.Autofac.DepdencyInjection.Samples.WebApi
             CreateWebHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .ConfigureServices(sp => sp.AddAutofac())
-                .UseStartup<Startup>();
+        private static IHostBuilder CreateWebHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .UseServiceProviderFactory(new AutofacServiceProviderFactory())
+                .ConfigureWebHostDefaults(builder => builder.UseStartup<Startup>());
     }
 }