+++ samples/MediatR.Extensions.Autofac.DependencyInjection.Samples.Shared/Commands/CustomerAddCommand.cs
@@ -4,7 +4,7 @@ namespace MediatR.Extensions.Autofac.DependencyInjection.Samples.Shared.Commands
 {
     public class CustomerAddCommand : IRequest
     {
-        public Guid Id { get;  }
+        public Guid Id { get; }
 
         public string Name { get; }