+++ src/MediatR.Extensions.Autofac.DependencyInjection/MediatRModule.cs
@@ -1,5 +1,10 @@
-﻿using System.Reflection;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
 using Autofac;
+using Autofac.Core;
+using Autofac.Features.Variance;
 using MediatR.Pipeline;
 using Module = Autofac.Module;
 
@@ -8,10 +13,12 @@ namespace MediatR.Extensions.Autofac.DependencyInjection
     internal class MediatRModule : Module
     {
         private readonly Assembly[] assemblies;
+        private readonly Type[] customBehaviorTypes;
 
-        public MediatRModule(Assembly[] assemblies)
+        public MediatRModule(Assembly[] assemblies, Type[] customBehaviorTypes)
         {
             this.assemblies = assemblies;
+            this.customBehaviorTypes = customBehaviorTypes;
         }
 
         protected override void Load(ContainerBuilder builder)
@@ -31,8 +38,15 @@ namespace MediatR.Extensions.Autofac.DependencyInjection
                     .AsClosedTypesOf(openHandlerType)
                     .AsImplementedInterfaces();
             }
-
+            
             builder.RegisterGeneric(typeof(RequestPostProcessorBehavior<,>)).As(typeof(IPipelineBehavior<,>));
+            
+            foreach (var customBehaviorType in this.customBehaviorTypes)
+            {
+                builder.RegisterGeneric(customBehaviorType)
+                    .As(typeof(IPipelineBehavior<,>));
+            }
+
             builder.RegisterGeneric(typeof(RequestPreProcessorBehavior<,>)).As(typeof(IPipelineBehavior<,>));
 
             builder.Register<ServiceFactory>(outerContext =>