+++ test/MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure/Behaviors/INoopRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure.Behaviors
+{
+    public interface INoopRequest<out TResponse> : IRequest<TResponse>
+    {
+        
+    }
+}