+++ test/MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure/Behaviors/UnconstraintedBehavior.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure.Behaviors
+{
+    public class UnconstraintedBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+    {
+        public Task<TResponse> Handle(TRequest request, CancellationToken cancellationToken, RequestHandlerDelegate<TResponse> next)
+        {
+            return next();
+        }
+    }
+}