+++ test/MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure/Commands/ResponseCommand.cs
@@ -1,7 +1,8 @@
-﻿namespace MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure.Commands
+﻿using MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure.Behaviors;
+
+namespace MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure.Commands
 {
-    public class ResponseCommand : IRequest<Response>
+    public class ResponseCommand : INoopRequest<Response>
     {
-        
     }
 }