+++ test/MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure/Commands/UnconstraintedCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace MediatR.Extensions.Autofac.DependencyInjection.TestInfrastructure.Commands
+{
+    public class UnconstraintedCommand : IRequest<int>
+    {
+        
+    }
+}